% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNCI_ses.R
\name{DNCI.ses}
\alias{DNCI.ses}
\title{DNCI Function : Dispersal-Niche Continuum Index}
\usage{
DNCI.ses(
  x,
  grouping,
  id = "no_name",
  Nperm = 1000,
  count = TRUE,
  plotSIMPER = TRUE
)
}
\arguments{
\item{x}{Sample/Taxa matrix with sample in row and taxa in column}

\item{grouping}{Grouping vector, ex : c(1,1,1,1,2,2,2,2,2) : 2 groups only !!}

\item{id}{Name of the dataset, default = "no_name"}

\item{Nperm}{Number of permutation, default = 1000, should be change to 100 for robustness analysis}

\item{count}{Display the number of permutation done, can be usefull with very large or small matrix, default = TRUE}

\item{plotSIMPER}{Display the SIMPER, PerSIMPER and E index plots, default = TRUE}

\item{dataType}{Need to be set for presence/absence or abundance data ("count"), default = "prab" (presence_absence)}
}
\description{
Quantitative identification of the main assembly process.
This function is based on PerSIMPER function and its E index return().
The three distributions of E index (corresponding to the three hypothesis: niche, dispersal, niche+dispersal) are used to compute the DNCI index.
If DNCI is significantly < 0 : dispersal || DNCI significantly > 0 : niche || DNCI +- CI ~~ 0 : dispersal+niche
See Vilmi et al. 2021 Ecography for DNCI computation and more information on process identification.
More information in code and comments inside function file.
}
\examples{
A <- DNCI.ses(Matrix, Group)
#where Matrix is a presence/absence matrix with taxa in column and sample in row
#and Group is a vector with length() == number of rows/samples in Matrix, 2 groups ONLY
#
B <- DNCI.ses(Matrix, Group, Nperm = 100, count = FALSE, plotSIMPER = FALSE)
#In this example, same data are analysed, with 100 permutations, with no countdown and no plots


}
