% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNCI_ses_overall.R
\name{DNCI.ses_overall}
\alias{DNCI.ses_overall}
\title{DNCI Function : Dispersal-Niche Continuum Index computation for overall dissimilarity analysis of 3 groups or more}
\usage{
DNCI.ses_overall(
  x,
  grouping,
  id = "no_name",
  Nperm = 1000,
  count = TRUE,
  dataTYPE = "prab",
  plotSIMPER = TRUE,
  parallelComputing = FALSE
)
}
\arguments{
\item{x}{Sample/Taxa matrix with sample in row and taxa in column}

\item{grouping}{Grouping vector, ex : c(1,1,1,1,2,2,2,2,2,3,3,3) : 3 groups or more}

\item{id}{Name of the dataset, default = "no_name"}

\item{Nperm}{Number of permutation, default = 1000, should be change to 100 for robustness analysis}

\item{count}{Display the number of permutation done, can be usefull with very large or small matrix, default = TRUE}

\item{dataTYPE}{Need to be set for presence/absence or abundance data ("count"), default = "prab" (presence_absence)}

\item{plotSIMPER}{Display the SIMPER, PerSIMPER and E index plots, default = TRUE}

\item{parallelComputing}{Run PerSIMPER on half of the available cores/nodes}
}
\description{
Quantitative identification of the main assembly process driving the overall dissimilarity contribution distribution (i.e. the empirical SIMPER profile structure).
This function is based on PerSIMPER_overall function and its E index return(). -Under development-
The three distributions of E index (corresponding to the three hypothesis: niche, dispersal, niche+dispersal) are used to compute the DNCI index.
If DNCI is significantly < 0 : dispersal || DNCI significantly > 0 : niche || DNCI +- CI ~~ 0 : dispersal+niche
-under development-
}
\examples{
A <- DNCImper:::DNCI.ses_overall(DNCImper::Matrix_4groups, DNCImper::Group4)
#where Matrix is a presence/absence matrix with taxa in column and sample in row
#and Group is a vector with length() == number of rows/samples in Matrix, 3 groups or more.
#
B <- DNCImper:::DNCI.ses_overall(DNCImper::Matrix_4groups, DNCImper::Group4, Nperm = 100, count = FALSE, plotSIMPER = FALSE)
#In this example, same data are analysed, with 100 permutations, with no countdown and no plots
}
