% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNCI_ses_overall_symmetrized.R
\name{DNCI.ses_overall_symmetrized}
\alias{DNCI.ses_overall_symmetrized}
\title{DNCI Function : Dispersal-Niche Continuum Index computation for overall dissimilarity analysis of 3 groups or more with randomly even groups}
\usage{
DNCI.ses_overall_symmetrized(
  Mat,
  Group,
  id = "no_name",
  NbrReRun = 100,
  Nperm = 100,
  count = FALSE
)
}
\arguments{
\item{id}{Name of the dataset, default = "no_name"}

\item{NbrReRun}{Number of iteration to obtain mean DNCI_overall values with even groups}

\item{Nperm}{Number of permutation, default = 1000, should be change to 100 for robustness analysis}

\item{count}{Display the number of permutation done, can be usefull with very large or small matrix, default = TRUE}

\item{x}{Sample/Taxa matrix with sample in row and taxa in column}

\item{grouping}{Grouping vector, ex : c(1,1,1,1,2,2,2,2,2,3,3,3) : 3 groups or more}

\item{dataType}{Need to be set for presence/absence or abundance data ("count"), default = "prab" (presence_absence)}

\item{plotSIMPER}{Display the SIMPER, PerSIMPER and E index plots, default = TRUE}
}
\description{
Quantitative identification of the main assembly process driving the overall dissimilarity contribution distribution (i.e. the empirical SIMPER profile structure).
This function is based on DNCI.ses_overall() PerSIMPER_overall function and its E index return(). -Under development-
-under development-
Group are made even by subsampling largest group to the size of the smallest ! CAUTION ! Do rerun for robust results
-under development-
}
\examples{
A <- DNCImper:::DNCI.ses_overall(Matrix, Group)
#where Matrix is a presence/absence matrix with taxa in column and sample in row
#and Group is a vector with length() == number of rows/samples in Matrix, 3 groups or more.
#
B <- DNCImper:::DNCI.ses_overall(Matrix, Group, Nperm = 100, count = FALSE, plotSIMPER = FALSE)
#In this example, same data are analysed, with 100 permutations, with no countdown and no plots


}
