% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Group_Tutorial.R
\docType{data}
\name{Group}
\alias{Group}
\title{Group vector made of 2 groups of assemblages}
\format{
An object of class data.frame()
}
\usage{
data(Data_Group_Tutorial)
}
\description{
Data from paleontological datasets: marine invertebrate from paleozoic
}
\examples{
data(Data_Group_Tutorial)
data(Data_Matrix_Tutorial)
PerSIMPER(Matrix, Group)

}
\references{
In prep.
}
\keyword{datasets,}
\keyword{matrix}
