% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Group4_Tutorial.R
\docType{data}
\name{Group4}
\alias{Group4}
\title{Group vector made of 4 groups of assemblages}
\format{
An object of class data.frame()
}
\usage{
data(Data_Group4_Tutorial)
}
\description{
Data from paleontological datasets: marine invertebrate from paleozoic
}
\examples{
data(Data_Group4_Tutorial)
data(Data_Matrix_Tutorial_4Groups)
DNCI_multigroup(Matrix_4groups, Group4)

}
\references{
In prep.
}
\keyword{datasets,}
\keyword{matrix}
