% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Matrix_Tutorial_4Groups.R
\docType{data}
\name{Matrix_4groups}
\alias{Matrix_4groups}
\title{Presence absence matrix made of 4 groups of assemblages}
\format{
An object of class data.frame()
}
\usage{
data(Data_Matrix_Tutorial_4Groups)
}
\description{
Data from paleontological datasets: marine invertebrate from paleozoic
}
\examples{
data(Data_Group4_Tutorial)
data(Data_Matrix_Tutorial_4Groups)
DNCI_multigroup(Matrix_4groups, Group4)

}
\references{
In prep.
}
\keyword{datasets,}
\keyword{matrix}
