% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerSIMPER.R
\name{PerSIMPER}
\alias{PerSIMPER}
\title{Per SIMPER function : identification of the main assembly process}
\usage{
PerSIMPER(
  matrixSIMP,
  Groups,
  leg = FALSE,
  count = TRUE,
  dataTYPE = "prab",
  Nperm = 1000,
  plotSIMPER = TRUE
)
}
\arguments{
\item{matrixSIMP}{Sample/Taxa matrix with sample in row and taxa in column}

\item{Groups}{Grouping vector, ex : c(1,1,1,1,2,2,2,2,2) : 2 groups only !!}

\item{leg}{Display the legend on PerSIMPER profile, default = FALSE}

\item{count}{Display the number of permutation done, can be usefull with very large or small matrix, default = TRUE}

\item{Nperm}{Number of permutation, default = 1000, should be change to 100 for robustness analysis}

\item{plotSIMPER}{Display the SIMPER, PerSIMPER and E index plots, default = TRUE}

\item{dataType}{Need to be set for presence/absence or abundance data ("count"), default = "prab" (presence_absence)}
}
\description{
This function is the basis of DNCImper package and DNCI analysis. It will permute the empirical matrix and produce the empirical as well as the randomized SIMPER profiles. Identify the main assembly process by comparing profiles. Permutations are fixed by rows, columns or both corresponding respectively, to niche, dispersal and niche+dispersal hypothesis.
The E index plot is produced to highlight the main assembly process. See Gibert & Escarguel 2019 Global Ecology and Biogeography for theory and more information on process identification.
More information in code and comments inside function file.
}
\examples{
A <- PerSIMPER(Matrix, Group)
#where Matrix is a presence/absence matrix with taxa in column and sample in row
#and Group is a vector with length() == number of rows/samples in Matrix, 2 groups ONLY
#
B <- PerSIMPER(Matrix, Group, Nperm = 100, count = FALSE, plotSIMPER = FALSE)
#In this example, same data are analysed, with 100 permutations, with no countdown and no plots


}
